# chore: change List<double> to Float32List for nearly 2x performance.