--- a/lib/array_math.dart
+++ b/lib/array_math.dart
@@ -1,26 +1,25 @@
+import 'dart:typed_data';
 
-
-
-
-
-
-
-
-void multiplyAdd({required double factorA, required List<double> factorB, required List<double> dest})
-{
+void multiplyAdd({
+  required double factorA,
+  required Float32List factorB,
+  required Float32List dest,
+}) {
   assert(factorB.length == dest.length);
 
-  for (var i = 0; i < dest.length; i++)
-  {
-      dest[i] += factorA * factorB[i];
+  for (var i = 0; i < dest.length; i++) {
+    dest[i] += factorA * factorB[i];
   }
 }
 
-void multiplyAddStep({required double factorA, required double step,  required List<double> factorB, required List<double> dest})
-{
-  for (var i = 0; i < dest.length; i++)
-  {
-      dest[i] += factorA * factorB[i];
-      factorA += step;
+void multiplyAddStep({
+  required double factorA,
+  required double step,
+  required Float32List factorB,
+  required Float32List dest,
+}) {
+  for (var i = 0; i < dest.length; i++) {
+    dest[i] += factorA * factorB[i];
+    factorA += step;
   }
 }