--- a/lib/audio_renderer_ex.dart
+++ b/lib/audio_renderer_ex.dart
@@ -1,166 +1,86 @@
-﻿
-
-
-
-
-
+﻿import 'dart:typed_data';
 
 import 'audio_renderer.dart';
-import 'array_int16.dart';
-
 
 /// Provides utility methods to convert the format of samples.
-extension AudioRenderEx on AudioRenderer
-{
-    /// <summary>
-    /// Renders the waveform as a stereo interleaved signal.
-    /// </summary>
-    /// <param name="renderer">The audio renderer.</param>
-    /// <param name="destination">The destination buffer.</param>
-    /// <remarks>
-    /// This utility method internally uses <see cref="ArrayPool{T}"/>,
-    /// which may result in memory allocation on the first call.
-    /// To completely avoid memory allocation,
-    /// use <see cref="IAudioRenderer.Render(Span{float}, Span{float})"/>.
-    /// </remarks>
-    void renderInterleaved(List<double> destination, {int offset = 0, int? length})
-    {
-        if (destination.length % 2 != 0)
-        {
-            throw "The length of the destination buffer must be even.";
-        }
-
-        int sampleCount = 0;
-
-        if (length != null) {
-          sampleCount = length;
-        } else {
-          sampleCount = destination.length ~/ 2;
-          sampleCount -= offset;
-        }
-
-        List<double> left = List<double>.filled(sampleCount, 0, growable: false);
-        List<double> right = List<double>.filled(sampleCount, 0, growable: false);
-
-        render(left, right);
-
-        for (var t = 0; t < sampleCount; t++)
-        {
-            destination[offset + t * 2 + 0] = left[t];
-            destination[offset + t * 2 + 1] = right[t];
-        }
+extension AudioRenderEx on AudioRenderer {
+  /// Renders the waveform as a stereo interleaved signal.
+  void renderInterleaved(
+    Float32List destination, {
+    int offset = 0,
+    int? length,
+  }) {
+    if (destination.length % 2 != 0) {
+      throw 'The length of the destination buffer must be even.';
     }
 
-    /// <summary>
-    /// Renders the waveform as a monaural signal.
-    /// </summary>
-    /// <param name="renderer">The audio renderer.</param>
-    /// <param name="destination">The destination buffer.</param>
-    /// <remarks>
-    /// This utility method internally uses <see cref="ArrayPool{T}"/>,
-    /// which may result in memory allocation on the first call.
-    /// To completely avoid memory allocation,
-    /// use <see cref="IAudioRenderer.Render(Span{float}, Span{float})"/>.
-    /// </remarks>
-    void renderMono(List<double> destination, {int offset = 0})
-    {
-        int sampleCount = destination.length ~/ 2;
+    int sampleCount = length ?? (destination.length ~/ 2 - offset);
 
-        sampleCount -= offset;
+    final left = Float32List(sampleCount);
+    final right = Float32List(sampleCount);
 
-        List<double> left = List<double>.filled(sampleCount, 0, growable: false);
-        List<double> right = List<double>.filled(sampleCount, 0, growable: false);
+    render(left, right);
 
-        render(left, right);
+    for (var t = 0; t < sampleCount; t++) {
+      destination[offset + t * 2] = left[t];
+      destination[offset + t * 2 + 1] = right[t];
+    }
+  }
 
-        for (var t = 0; t < sampleCount; t++)
-        {
-            destination[offset + t] = (left[t] + right[t]) / 2;
-        }
+  /// Renders the waveform as a monaural signal.
+  void renderMono(Float32List destination, {int offset = 0}) {
+    int sampleCount = destination.length - offset;
+
+    final left = Float32List(sampleCount);
+    final right = Float32List(sampleCount);
+
+    render(left, right);
+
+    for (var t = 0; t < sampleCount; t++) {
+      destination[offset + t] = (left[t] + right[t]) / 2;
+    }
+  }
+
+  /// Renders the waveform as a stereo interleaved signal with 16-bit quantization.
+  void renderInterleavedInt16(
+    Int16List destination, {
+    int offset = 0,
+    int? length,
+  }) {
+    if (destination.length % 2 != 0) {
+      throw 'Invalid destination length';
     }
 
-    /// <summary>
-    /// Renders the waveform as a stereo interleaved signal with 16-bit quantization.
-    /// </summary>
-    /// <param name="renderer">The audio renderer.</param>
-    /// <param name="destination">The destination buffer.</param>
-    /// <remarks>
-    /// Out of range samples will be clipped.
-    /// This utility method internally uses <see cref="ArrayPool{T}"/>,
-    /// which may result in memory allocation on the first call.
-    /// To completely avoid memory allocation,
-    /// use <see cref="IAudioRenderer.Render(Span{float}, Span{float})"/>.
-    /// </remarks>
-    void renderInterleavedInt16(ArrayInt16 destination, {int offset = 0, int? length})
-    {
-        if (destination.bytes.lengthInBytes % 4 != 0)
-        {
-            throw "Invalid destination length";
-        }
+    int sampleCount = length ?? (destination.length ~/ 2 - offset);
 
-        int sampleCount = 0;
+    final left = Float32List(sampleCount);
+    final right = Float32List(sampleCount);
 
-        if (length != null) {
-          sampleCount = length;
-        } else {
-          sampleCount = destination.bytes.lengthInBytes ~/ 4;
-          sampleCount -= offset;
-        }
+    render(left, right);
 
-        List<double> left = List<double>.filled(sampleCount, 0, growable: false);
-        List<double> right = List<double>.filled(sampleCount, 0, growable: false);
+    for (var t = 0; t < sampleCount; t++) {
+      // Clamp values between -32768 and 32767 to prevent overflow
+      int sampleLeft = (32768 * left[t]).clamp(-32768, 32767).toInt();
+      int sampleRight = (32768 * right[t]).clamp(-32768, 32767).toInt();
 
-        render(left, right);
-
-        for (var t = 0; t < sampleCount; t++)
-        {
-            int sampleLeft = (32768 * left[t]).toInt();
-            int sampleRight = (32768 * right[t]).toInt();
-
-            // these get automaticall casted to shorts in ArrayInt16[]
-            destination[offset + t * 2 + 0] = sampleLeft;
-            destination[offset + t * 2 + 1] = sampleRight;
-        }
+      destination[offset + t * 2] = sampleLeft;
+      destination[offset + t * 2 + 1] = sampleRight;
     }
+  }
 
-    /// <summary>
-    /// Renders the waveform as a monaural signal with 16-bit quantization.
-    /// </summary>
-    /// <param name="renderer">The audio renderer.</param>
-    /// <param name="destination">The destination buffer.</param>
-    /// <remarks>
-    /// Out of range samples will be clipped.
-    /// This utility method internally uses <see cref="ArrayPool{T}"/>,
-    /// which may result in memory allocation on the first call.
-    /// To completely avoid memory allocation,
-    /// use <see cref="IAudioRenderer.Render(Span{float}, Span{float})"/>.
-    /// </remarks>
-    void renderMonoInt16(ArrayInt16 destination, {int offset = 0, int? length})
-    {
-        if (destination.bytes.lengthInBytes % 2 != 0)
-        {
-            throw "Invalid destination length";
-        }
+  /// Renders the waveform as a monaural signal with 16-bit quantization.
+  void renderMonoInt16(Int16List destination, {int offset = 0, int? length}) {
+    int sampleCount = length ?? (destination.length - offset);
 
-        int sampleCount = 0;
+    final left = Float32List(sampleCount);
+    final right = Float32List(sampleCount);
 
-        if (length != null) {
-          sampleCount = length;
-        } else {
-          sampleCount = destination.bytes.lengthInBytes ~/ 2;
-          sampleCount -= offset;
-        }
+    render(left, right);
 
-        List<double> left = List<double>.filled(sampleCount, 0, growable: false);
-        List<double> right = List<double>.filled(sampleCount, 0, growable: false);
-
-        render(left, right);
-
-        for (var t = 0; t < sampleCount; t++)
-        {
-            int sample = (16384 * (left[t] + right[t])).toInt();
-            destination[offset + t] = sample;
-        }
+    for (var t = 0; t < sampleCount; t++) {
+      // Mix to mono and convert to 16-bit integer with clamping
+      int sample = (16384 * (left[t] + right[t])).clamp(-32768, 32767).toInt();
+      destination[offset + t] = sample;
     }
+  }
 }
-