--- a/lib/dart_melty_soundfont.dart
+++ b/lib/dart_melty_soundfont.dart
@@ -1,8 +1,8 @@
-export 'synthesizer.dart';
-export 'synthesizer_settings.dart';
+export 'dart:typed_data';
+
+export 'audio_renderer_ex.dart';
 export 'midi_file.dart';
 export 'midi_file_sequencer.dart';
 export 'preset.dart';
-export 'audio_renderer_ex.dart';
-export 'array_int16.dart';
-export 'dart:typed_data';
+export 'synthesizer.dart';
+export 'synthesizer_settings.dart';