# A total overhaul + 2x performance

## The proof

<img width="1512" alt="Screenshot 2025-02-25 at 13 07 10" src="https://github.com/user-attachments/assets/7d1d1bcd-e5e4-4fd0-a907-80db5ba4a79b" />

On the RIGHT, you can see the new performance in the terminal. (6s & 12s)
On the LEFT, you can see the old performance in the terminal. (10s & 20s)
Smaller is better ofc, in this case.

They both use the same benchmark test of 1 million renders. Tested with bigger and smaller block sizes as displayed in the bench mark logs. The benchmark can be seen in the image as well. They are the same.

## Changes
- All files have been formatted with the standard dart formatter.
- All occurances of List<double> have been changed to Float32List.
- ArrayInt16 (the custom class) was removed and Int16List replaces it.
- Added the benchmark file to the root for good measure.
- Updated the example to reflect the changes.

I have tried this in my real app, and it works brilliantly well!

Let me know what you think 🙏 

Ps. I got the inspiration from this [video](https://www.youtube.com/watch?v=HF8gHAakb1Q) which renders an image with dart and compares it to rust and c++. Dart gets nearly the same preformance in that particular case. I just thought I'd use the same principle in this.

Pps. I decided to format all the files, because they were a bit all over the place.