@saarikivit 2025-02-25
☝️ I removed the linter based cleanup as it had errors in my other applications, and it was unnecessary.